fix: fixing flakeyness in tests
