+++ e2e/config_repositories_test.go
@@ -1,39 +1,58 @@
-package e2e
+package e2e_test
 
 import (
+	"encoding/json"
+	"slices"
 	"testing"
 
+	"github.com/cycloidio/cycloid-cli/client/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConfigRepositories(t *testing.T) {
+
+	var (
+		CRUrl       = "git@github.com:cycloidio/cycloid-cli-test-catalog.git"
+		CRBranch    = "config-e2e"
+		CRCanonical = "config-repo-e2e"
+		CRCred      = config.ConfigRepo.CredentialCanonical
+	)
+
 	defer t.Run("SuccessConfigRepositoriesDelete", func(t *testing.T) {
 		cmdOut, cmdErr := executeCommand([]string{
 			"--output", "json",
 			"--org", config.Org,
 			"config-repo",
 			"delete",
-			"--canonical", "test-config",
+			"--canonical", CRCanonical,
 		})
 
 		assert.Nil(t, cmdErr)
-		require.Equal(t, "", string(cmdOut))
+		require.Empty(t, cmdOut)
 	})
+
 	t.Run("SuccessConfigRepositoriesCreate", func(t *testing.T) {
 		cmdOut, cmdErr := executeCommand([]string{
 			"--output", "json",
 			"--org", config.Org,
 			"config-repo",
 			"create",
-			"--name", "test-config",
-			"--branch", "",
-			"--cred", config.ConfigRepo.CredentialCanonical,
-			"--url", *config.ConfigRepo.Canonical,
+			"--name", CRCanonical,
+			"--branch", CRBranch,
+			"--cred", CRCanonical,
+			"--url", CRUrl,
 		})
 
 		assert.Nil(t, cmdErr)
-		require.Contains(t, cmdOut, "canonical\": \"test-config")
+
+		var result models.ConfigRepository
+		err := json.Unmarshal([]byte(cmdOut), &result)
+		assert.NoError(t, err)
+		assert.Equal(t, CRUrl, *result.URL)
+		assert.Equal(t, CRBranch, result.Branch)
+		assert.Equal(t, CRCanonical, *result.Canonical)
+		assert.Equal(t, CRCred, result.CredentialCanonical)
 	})
 
 	t.Run("SuccessConfigRepositoriesList", func(t *testing.T) {
@@ -45,7 +64,21 @@ func TestConfigRepositories(t *testing.T) {
 		})
 
 		assert.Nil(t, cmdErr)
-		require.Contains(t, cmdOut, "canonical\": \"test-config")
+		var list []models.ConfigRepository
+		err := json.Unmarshal([]byte(cmdOut), &list)
+		assert.NoError(t, err)
+
+		if index := slices.IndexFunc(list, func(cr models.ConfigRepository) bool {
+			return *cr.Canonical == CRCanonical
+		}); index != -1 {
+			result := list[index]
+			assert.Equal(t, CRUrl, *result.URL)
+			assert.Equal(t, CRBranch, result.Branch)
+			assert.Equal(t, CRCanonical, *result.Canonical)
+			assert.Equal(t, CRCred, result.CredentialCanonical)
+		} else {
+			t.Fatal("did not found our config repo in cmd output: " + cmdOut)
+		}
 	})
 
 	t.Run("SuccessConfigRepositoriesGet", func(t *testing.T) {
@@ -54,11 +87,17 @@ func TestConfigRepositories(t *testing.T) {
 			"--org", config.Org,
 			"config-repo",
 			"get",
-			"--canonical", "test-config",
+			"--canonical", CRCanonical,
 		})
 
 		assert.Nil(t, cmdErr)
-		require.Contains(t, cmdOut, "canonical\": \"test-config")
-	})
 
+		var result models.ConfigRepository
+		err := json.Unmarshal([]byte(cmdOut), &result)
+		assert.NoError(t, err)
+		assert.Equal(t, CRUrl, *result.URL)
+		assert.Equal(t, CRBranch, result.Branch)
+		assert.Equal(t, CRCanonical, *result.Canonical)
+		assert.Equal(t, CRCred, result.CredentialCanonical)
+	})
 }
