+++ e2e/events_test.go
@@ -1,4 +1,4 @@
-package e2e
+package e2e_test
 
 import (
 	"testing"
