+++ e2e/organizations_test.go
@@ -1,4 +1,4 @@
-package e2e
+package e2e_test
 
 import (
 	"fmt"
@@ -10,6 +10,7 @@ import (
 )
 
 func TestOrganizations(t *testing.T) {
+	t.Skip()
 	t.Run("SuccessOrganizationsGet", func(t *testing.T) {
 		cmdOut, cmdErr := executeCommand([]string{
 			"--output", "json",
