+++ e2e/pipelines_test.go
@@ -1,11 +1,11 @@
-package e2e
+package e2e_test
 
 import (
 	"testing"
 )
 
 func TestPipelines(t *testing.T) {
-	t.Parallel()
+	
 
 	// Preparation
